+++ src/Paket/Program.fs
@@ -8,28 +8,33 @@ type Command =
     | Install
     | Update
     | Outdated
+    | ConvertFromNuget
     | Unknown
 
 type CLIArguments =
     | [<First>][<NoAppSettings>][<CustomCommandLine("install")>] Install
     | [<First>][<NoAppSettings>][<CustomCommandLine("update")>] Update
     | [<First>][<NoAppSettings>][<CustomCommandLine("outdated")>] Outdated
+    | [<First>][<NoAppSettings>][<CustomCommandLine("convert-from-nuget")>] ConvertFromNuget
     | [<AltCommandLine("-v")>] Verbose
-    | Dependencies_File of string
+    | Dependencies_file of string
     | [<AltCommandLine("-f")>] Force
+    | Hard
 with
     interface IArgParserTemplate with
         member s.Usage =
             match s with
             | Install -> "installs all packages."
-            | Update -> "updates the Lock File and installs all packages."
+            | Update -> "updates the packet.lock ile and installs all packages."
             | Outdated -> "displays information about new packages."
+            | ConvertFromNuget -> "converts all projects from NuGet to Paket."
             | Verbose -> "displays verbose output."
-            | Dependencies_File _ -> "specify a file containing dependency definitions."
-            | Force -> "specify a dependency definition."
+            | Dependencies_file _ -> "specify a file containing dependency definitions."
+            | Force -> "forces the download of all packages."
+            | Hard -> "overwrites manual package references."
 
 
-let parser = UnionArgParser.Create<CLIArguments>("USAGE: paket [install|update|outdated] ... options")
+let parser = UnionArgParser.Create<CLIArguments>("USAGE: paket [install|update|outdated|convert-from-nuget] ... options")
  
 let results,verbose =
     try
@@ -38,6 +43,7 @@ let results,verbose =
             if results.Contains <@ CLIArguments.Install @> then Command.Install
             elif results.Contains <@ CLIArguments.Update @> then Command.Update
             elif results.Contains <@ CLIArguments.Outdated @> then Command.Outdated
+            elif results.Contains <@ CLIArguments.ConvertFromNuget @> then Command.ConvertFromNuget
             else Command.Unknown
         Some(command,results),results.Contains <@ CLIArguments.Verbose @>
     with
@@ -49,7 +55,7 @@ try
     match results with
     | Some(command,results) ->
         let dependenciesFile = 
-            match results.TryGetResult <@ CLIArguments.Dependencies_File @> with
+            match results.TryGetResult <@ CLIArguments.Dependencies_file @> with
             | Some x -> x
             | _ -> "paket.dependencies"
 
@@ -58,10 +64,16 @@ try
             | Some _ -> true
             | None -> false
 
+        let hard = 
+            match results.TryGetResult <@ CLIArguments.Hard @> with
+            | Some _ -> true
+            | None -> false
+
         match command with
-        | Command.Install -> Process.Install(false,force,dependenciesFile)
-        | Command.Update -> Process.Install(true,force,dependenciesFile)
+        | Command.Install -> Process.Install(false,force,hard,dependenciesFile)
+        | Command.Update -> Process.Install(true,force,hard,dependenciesFile)
         | Command.Outdated -> Process.ListOutdated(dependenciesFile)
+        | Command.ConvertFromNuget -> Process.ConvertFromNuget()
         | _ -> failwithf "no command given.%s" (parser.Usage())
         
         tracefn "Ready."