+++ src/Paket/SolutionFile.fs
@@ -1,0 +1,26 @@
+﻿namespace Paket
+
+open System.IO
+
+/// Contains methods to read and manipulate solution files.
+module SolutionFile =
+    let private removeNugetSlnFolderIfEmpty(slnContent: ResizeArray<string>) =
+        match slnContent |> Seq.tryFindIndex (fun line -> 
+                line.StartsWith("""Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}")""") && line.Contains(".nuget")) with
+        | Some(index) -> 
+            if slnContent.[index+1].Contains("ProjectSection(SolutionItems)") &&
+               slnContent.[index+2].Contains("EndProjectSection") &&
+               slnContent.[index+3].Contains("EndProject")
+            then 
+                slnContent.RemoveRange(index, 4)
+        | None -> ()
+
+    
+    let RemoveNugetPackagesFile(fileName: string) = 
+        let slnContent = ResizeArray( File.ReadAllLines fileName )
+        match slnContent |> Seq.tryFindIndex (fun line -> line.Contains(".nuget\packages.config")) with
+        | Some(index) -> 
+            slnContent.RemoveAt(index)
+            removeNugetSlnFolderIfEmpty(slnContent)
+            File.WriteAllLines(fileName, slnContent)
+        | None -> ()