+++ RELEASE_NOTES.md
@@ -1,5 +1,14 @@
-#### 0.1.1 - 12.09.2014
-* We are live - yay!!
+#### 0.2.0-alpha001 - 16.09.2014
+* convert-from-nuget command - http://fsprojects.github.io/Paket/convert_from_nuget.html
+
+#### 0.1.4 - 16.09.2014
+* Only vbproj, csproj and fsproj files are handled
+
+#### 0.1.3 - 15.09.2014
+* Detect FSharpx.Core in packages
+
+#### 0.1.2 - 15.09.2014
+* --hard parameter allows better transition from NuGet.exe
 
 #### 0.1.0 - 12.09.2014
 * We are live - yay!